pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hatelens"
version = "0.4.0"
description = "Tiny LLMs + LoRA for efficient, explainable hate speech detection (DynaHate, HateCheck)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
  { name = "Vittoria Meroni" },
  { name = "Emanuele Rimoldi" },
  { name = "Simone Vicentini" },
]
keywords = ["hate-speech", "explainability", "LoRA", "TinyLLM", "NLP"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "torch>=2.2.0",
  "transformers>=4.40.0",
  "peft>=0.11.0",
  "datasets>=2.16.0",
  "evaluate>=0.4.0",
  "accelerate>=0.27.0",
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "scikit-learn>=1.3.0",
  "PyYAML>=6.0",
  "tqdm>=4.66.0",
  "matplotlib>=3.8.0",
]

[project.optional-dependencies]
lime = ["lime>=0.2.0.1"]
wandb = ["wandb>=0.16.0"]
quant = ["bitsandbytes>=0.43.0"]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
]
all = ["lime>=0.2.0.1", "wandb>=0.16.0", "bitsandbytes>=0.43.0"]

[project.scripts]
hatelens = "hatelens.cli:main"

[project.urls]
Homepage = "https://github.com/EmaRimoldi/HateLens"
Repository = "https://github.com/EmaRimoldi/HateLens"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
